+++ coco/signingMessages_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/dedis/crypto/nist"
-	"github.com/dedis/prifi/coconet"
+	"github.com/dedis/prifi/coco/coconet"
 	"github.com/dedis/prifi/timestamp"
 )
