+++ coco/signingNode.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/dedis/crypto/abstract"
-	"github.com/dedis/prifi/coconet"
+	"github.com/dedis/prifi/coco/coconet"
 	"github.com/dedis/prifi/timestamp"
 )
